LayoutTests/editing/undomanager/undomanager-item.html

<!DOCTYPE html>
<html>
<head>
<title>undoManager.item()</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
This tests UndoManager's item() method.

<div id="edit" contentEditable>Hello</div>

<script>
var undoManager = document.undoManager;

test(function() {
    assert_idl_attribute(undoManager, "item");
}, "undoManager has item property.");

var transaction = {"executeAutomatic": function() { }};
var transaction2 = {"executeAutomatic": function() { }};

test(function() {
    undoManager.transact(transaction);
    undoManager.transact(transaction2, true);
    
    var entry = undoManager.item(0);
    assert_equals(entry.length, 2);
    assert_equals(entry[0], transaction);
    assert_equals(entry[1], transaction2);
}, "After two transact calls that are merged into one, item(0) returns an array of "
    + "the two transactions in the correct order.");

test(function() {
    var entry = undoManager.item(1);
    assert_equals(entry, null);
}, "item(1) returns null because it's out of range.");

test(function() {
    undoManager.undo();

    var entry = undoManager.item(0);
    assert_equals(entry.length, 2);
    assert_equals(entry[0], transaction);
    assert_equals(entry[1], transaction2);
}, "After an undo call, item(0) stills return an array of the two transactions in the correct order.");

test(function() {
    undoManager.redo();
    assert_equals(undoManager.position, 0);
}, "After a redo call, undoManager's position is back to 0.");

test(function() {
    var range = document.createRange();
    range.selectNodeContents(document.getElementById("edit"));
    window.getSelection().addRange(range);
    document.execCommand("Bold");
    
    var entry = undoManager.item(0);
    assert_equals(entry.length, 1);
    assert_own_property(entry[0], "label");
}, "After executing an editing command, item(0) returns an array of an object with label property.");

test(function() {
    var entry = undoManager.item(1);
    assert_equals(entry.length, 2);
    assert_equals(entry[0], transaction);
    assert_equals(entry[1], transaction2);
}, "item(1) now returns an array of the two transactions.");

test(function() {
    assert_not_equals(undoManager.item(1), undoManager.item(1));
}, "item(1) returns a different object than item(1).");

document.getElementById("edit").style.display = "none";
</script>
</body>
</html>
